backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.20
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
bcrypt>=4.0.0
PyJWT>=2.8.0
openai>=1.0.0
replicate>=0.25.0
moviepy>=1.0.3
opencv-python>=4.8.0
pillow>=10.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0,<2.0
requests>=2.31.0
redis>=5.0.0
pyyaml>=6.0.0
# VBench evaluation library (when available)
# vbench>=0.1.0  # TODO: Install from GitHub: Vchitect/VBench when library becomes available

